Create LICENSE
We don't have a license on this repository yet so I wanted to add one that permits lots of sharing and reuse. Since you're the primary author on the materials @kgarofali, I wanted to submit this license suggestion as a PR for you to merge. How do you feel about the [MIT license](https://en.wikipedia.org/wiki/MIT_License)?